% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cwrite}
\alias{cwrite}
\title{cwrite: Write vector to file}
\usage{
cwrite(y, l, file)
}
\arguments{
\item{y}{vector to write.}

\item{l}{length of the vector; (you can provide length(y)).}

\item{file}{output filename.}
}
\description{
Write a vector to a file using C for fast writing. Not much more complicated than that. It's faster than the base R write functions.
}
\examples{
d = rnorm(1000);
cwrite(d, length(d), "output.txt")
}

