% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{CWriteR}
\alias{CWriteR}
\alias{CWriteR-package}
\title{Fast file writing with CWriteR.}
\description{
CWriteR provides functions that use C code to write
output files. These C functions are much faster than the
base R output functions, which is useful if you need to write
millions or billions of entries to a file quickly.
}
\author{
Nathan Sheffield
}
\references{
\url{http://www.github.com/sheffien}
}

